feat: creating exception handling

+++ src/application/useCases/beers/delete-beer.ts
@@ -1,11 +1,18 @@
 import { BeerRepository } from '@application/repositories/beers-repository';
 import { Injectable } from '@nestjs/common';
+import { BeerNotFoundException } from './errors/beer-not-found';
 
 @Injectable()
 export class DeleteBeer {
   constructor(private readonly beerRepository: BeerRepository) {}
 
   async execute(beerId: string) {
+    const beer = await this.beerRepository.listById(beerId);
+
+    if (!beer) {
+      throw new BeerNotFoundException();
+    }
+
     await this.beerRepository.delete(beerId);
   }
 }
